+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # [Versions](https://github.com/Tracktor/design-system/releases)
 
 ## v3.5.0
-- **[feat]** -`Alert` add secondary variant
+- **[feat]** -`Alert` add primary and secondary severity
 - **[fix]** -`Paper` border radius mobile